addressing specs
